+++ forge/src/main/java/dev/aika/smsn/forge/mixin/MixinInfo.java
@@ -19,6 +19,7 @@ public class MixinInfo {
                 "dev.aika.smsn.forge.mixin.quark.ThreadContributorListLoaderMixin", "quark",
                 "dev.aika.smsn.forge.mixin.quark.QButtonMixin", "quark",
                 "dev.aika.smsn.forge.mixin.titanium.RewardMixin", "titanium",
+                "dev.aika.smsn.forge.mixin.voidscape.DonatorHandlerMixin", "voidscape",
                 "dev.aika.smsn.forge.mixin.xaero.minimap.InternetMixin", "xaerominimap",
                 "dev.aika.smsn.forge.mixin.xaero.minimap.PatreonMixin", "xaerominimap",
                 "dev.aika.smsn.forge.mixin.xaero.world_map.InternetMixin", "xaeroworldmap",